test: hoist the boundary-crossing counter into backend_jit_helpers
Self-audit: the guard added in the previous commit grew a local _count_coercions,
and the vertical branch was about to add a near-identical _count_vertical_coercions.
Two copies of a general-purpose probe in two specific test modules is exactly the
duplication the audit rule forbids, so it goes in the shared helper module the
test_backend_* files already import from.

The docstring carries the rule that makes the check non-vacuous: always pair the
zero-crossing assertion with a negative control that is EXPECTED to cross.
